# loms/api/openapiv2/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Logistics Order Management System",
    "description": "API for managing orders in the Logistics Order Management System",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "OrdersService"
    },
    {
      "name": "StocksService"
    }
  ],
  "host": "localhost:8084",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/order/cancel": {
      "post": {
        "operationId": "OrdersService_CancelOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CancelOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CancelOrderRequest"
            }
          }
        ],
        "tags": [
          "OrdersService"
        ]
      }
    },
    "/order/create": {
      "post": {
        "operationId": "OrdersService_CreateOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateOrderRequest"
            }
          }
        ],
        "tags": [
          "OrdersService"
        ]
      }
    },
    "/order/info": {
      "get": {
        "operationId": "OrdersService_OrderInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1OrderInfoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "orderId",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "OrdersService"
        ]
      }
    },
    "/order/pay": {
      "post": {
        "operationId": "OrdersService_PayOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1PayOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1PayOrderRequest"
            }
          }
        ],
        "tags": [
          "OrdersService"
        ]
      }
    },
    "/stock/info": {
      "get": {
        "operationId": "StocksService_StocksInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1StocksInfoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "sku",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "StocksService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1CancelOrderRequest": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1CancelOrderResponse": {
      "type": "object"
    },
    "v1CreateOrderRequest": {
      "type": "object",
      "properties": {
        "user": {
          "type": "string",
          "format": "int64"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1OrderItem"
          }
        }
      }
    },
    "v1CreateOrderResponse": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1OrderInfoResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "user": {
          "type": "string",
          "format": "int64"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1OrderItem"
          }
        }
      }
    },
    "v1OrderItem": {
      "type": "object",
      "properties": {
        "sku": {
          "type": "string",
          "format": "int64"
        },
        "count": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "v1PayOrderRequest": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1PayOrderResponse": {
      "type": "object"
    },
    "v1StocksInfoResponse": {
      "type": "object",
      "properties": {
        "count": {
          "type": "integer",
          "format": "int64"
        }
      }
    }
  }
}